# whisper : add support for backends with multiple ggml_backend_buffer_type