--- a/src/whisper-arch.h
+++ b/src/whisper-arch.h
@@ -1,7 +1,3 @@
-// SPDX-FileCopyrightText: Copyright 2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
-// SPDX-License-Identifier: MIT
-//
-
 #pragma once
 
 #include "ggml.h"