--- a/src/whisper.cpp
+++ b/src/whisper.cpp
@@ -19,6 +19,7 @@
 #include <cassert>
 #define _USE_MATH_DEFINES
 #include <cmath>
+#include <climits>
 #include <codecvt>
 #include <cstdarg>
 #include <cstdio>