@bradmurray-dt 2025-03-11
Is anything additional needed to run this?  Do you have any performance comparisons?