@eddnjjn 2025-03-19
Just want to add that I’ve tested the patch using whisper-cli and whisper-bench in the following environments

Linux x86 (Ubuntu 20.04.6) – CPU (aarch64) backend
Android Pixel 8 – CPU (aarch64, KleidiAI) backend
macOS – Metal and CPU (with/without BLAS, aarch64, KleidiAI) backends