@ggerganov 2025-03-20
Thanks for the update. I'll do some testing soon on my devices and if everything looks OK, will merge.